server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-rest-framework-simplejwt-blacklist==1.0.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
reportlab==4.0.7
dj-database-url==2.1.0

